+++ glue/viewers/image/layer_artist.py
@@ -180,7 +180,9 @@ class ImageLayerArtist(BaseImageLayerArtist):
         if self.uuid is None or self.state.attribute is None or self.state.layer is None:
             return
 
-        changed = set() if force else self.pop_changed_properties()
+        # NOTE: we need to evaluate this even if force=True so that the cache
+        # of updated properties is up to date after this method has been called.
+        changed = self.pop_changed_properties()
 
         if force or any(prop in changed for prop in ('layer', 'attribute',
                                                      'slices', 'x_att', 'y_att')):
@@ -339,7 +341,9 @@ class ImageSubsetLayerArtist(BaseImageLayerArtist):
         if self.state.layer is None:
             return
 
-        changed = set() if force else self.pop_changed_properties()
+        # NOTE: we need to evaluate this even if force=True so that the cache
+        # of updated properties is up to date after this method has been called.
+        changed = self.pop_changed_properties()
 
         if force or any(prop in changed for prop in ('layer', 'attribute', 'color',
                                                      'x_att', 'y_att', 'slices')):