+++ glue/viewers/scatter/layer_artist.py
@@ -511,7 +511,9 @@ class ScatterLayerArtist(MatplotlibLayerArtist):
                 self.state.layer is None):
             return
 
-        changed = set() if force else self.pop_changed_properties()
+        # NOTE: we need to evaluate this even if force=True so that the cache
+        # of updated properties is up to date after this method has been called.
+        changed = self.pop_changed_properties()
 
         if force or len(changed & DATA_PROPERTIES) > 0:
             self._update_data()